configure sonar by default. move to own config file
